Enum etherparse::IpTrafficClass

pub enum IpTrafficClass {
    IPv6HeaderHopByHop,
    Icmp,
    Igmp,
    Ggp,
    IPv4,
    Stream,
    Tcp,
    Cbt,
    Egp,
    Igp,
    BbnRccMon,
    NvpII,
    Pup,
    Argus,
    Emcon,
    Xnet,
    Chaos,
    Udp,
    Mux,
    DcnMeas,
    Hmp,
    Prm,
    XnsIdp,
    Trunk1,
    Trunk2,
    Leaf1,
    Leaf2,
    Rdp,
    Irtp,
    IsoTp4,
    NetBlt,
    MfeNsp,
    MeritInp,
    Dccp,
    ThirdPartyConnectProtocol,
    Idpr,
    Xtp,
    Ddp,
    IdprCmtp,
    TpPlusPlus,
    Il,
    Ipv6,
    Sdrp,
    IPv6RouteHeader,
    IPv6FragmentationHeader,
    Idrp,
    Rsvp,
    Gre,
    Dsr,
    Bna,
    IPv6EncapSecurityPayload,
    IPv6AuthenticationHeader,
    Inlsp,
    Swipe,
    Narp,
    Mobile,
    Tlsp,
    Skip,
    IPv6Icmp,
    IPv6NoNextHeader,
    IPv6DestinationOptions,
    AnyHostInternalProtocol,
    Cftp,
    AnyLocalNetwork,
    SatExpak,
    Krytolan,
    Rvd,
    Ippc,
    AnyDistributedFileSystem,
    SatMon,
    Visa,
    Ipcv,
    Cpnx,
    Cphb,
    Wsn,
    Pvp,
    BrSatMon,
    SunNd,
    WbMon,
    WbExpak,
    IsoIp,
    Vmtp,
    SecureVmtp,
    Vines,
    TtpOrIptm,
    NsfnetIgp,
    Dgp,
    Tcf,
    Eigrp,
    Ospfigp,
    SpriteRpc,
    Larp,
    Mtp,
    Ax25,
    Ipip,
    Micp,
    SccSp,
    EtherIp,
    Encap,
    Gmtp,
    Ifmp,
    Pnni,
    Pim,
    Aris,
    Scps,
    Qnx,
    ActiveNetworks,
    IpComp,
    SitraNetworksProtocol,
    CompaqPeer,
    IpxInIp,
    Vrrp,
    Pgm,
    AnyZeroHopProtocol,
    Layer2TunnelingProtocol,
    Ddx,
    Iatp,
    Stp,
    Srp,
    Uti,
    SimpleMessageProtocol,
    Sm,
    Ptp,
    IsisOverIpv4,
    Fire,
    Crtp,
    Crudp,
    Sscopmce,
    Iplt,
    Sps,
    Pipe,
    Sctp,
    Fc,
    RsvpE2eIgnore,
    MobilityHeader,
    UdpLite,
    MplsInIp,
    Manet,
    Hip,
    Shim6,
    Wesp,
    Rohc,
    ExperimentalAndTesting0,
    ExperimentalAndTesting1,
}

Identifiers for the traffic_class field in ipv6 headers and protocol field in ipv4 headers.

Variants

IPv6HeaderHopByHop

IPv6 Hop-by-Hop Option [RFC8200]

Icmp

Internet Control Message [RFC792]

Igmp

Internet Group Management [RFC1112]

Ggp

Gateway-to-Gateway [RFC823]

IPv4

IPv4 encapsulation [RFC2003]

Stream

Stream [RFC1190][RFC1819]

Tcp

Transmission Control [RFC793]

Cbt

CBT [Tony_Ballardie]

Egp

Exterior Gateway Protocol [RFC888][David_Mills]

Igp

any private interior gateway (used by Cisco for their IGRP) [Internet_Assigned_Numbers_Authority]

BbnRccMon

BBN RCC Monitoring [Steve_Chipman]

NvpII

Network Voice Protocol [RFC741][Steve_Casner]

Pup

PUP

Argus

ARGUS (deprecated) [Robert_W_Scheifler]

Emcon

EMCON []

Xnet

Cross Net Debugger [Haverty, J., "XNET Formats for Internet Protocol Version 4", IEN 158, October 1980.][Jack_Haverty]

Chaos

Chaos [J_Noel_Chiappa]

Udp

User Datagram [RFC768][Jon_Postel]

Mux

Multiplexing [Cohen, D. and J. Postel, "Multiplexing Protocol", IEN 90, USC/Information Sciences Institute, May 1979.][Jon_Postel]

DcnMeas

DCN Measurement Subsystems [David_Mills]

Hmp

Host Monitoring [RFC869][Bob_Hinden]

Prm

Packet Radio Measurement [Zaw_Sing_Su]

XnsIdp

XEROX NS IDP

Trunk1

Trunk-1 [Barry_Boehm]

Trunk2

Trunk-2 [Barry_Boehm]

Leaf1

Leaf-1 [Barry_Boehm]

Leaf2

Leaf-2 [Barry_Boehm]

Rdp

Reliable Data Protocol [RFC908][Bob_Hinden]

Irtp

Internet Reliable Transaction [RFC938][Trudy_Miller]

IsoTp4

ISO Transport Protocol Class 4 [RFC905][]

NetBlt

Bulk Data Transfer Protocol [RFC969][David_Clark]

MfeNsp

MFE Network Services Protocol [Shuttleworth, B., "A Documentary of MFENet, a National Computer Network", UCRL-52317, Lawrence Livermore Labs, Livermore, California, June 1977.][Barry_Howard]

MeritInp

MERIT Internodal Protocol [Hans_Werner_Braun]

Dccp

Datagram Congestion Control Protocol [RFC4340]

ThirdPartyConnectProtocol

Third Party Connect Protocol [Stuart_A_Friedberg]

Idpr

Inter-Domain Policy Routing Protocol [Martha_Steenstrup]

Xtp

XTP [Greg_Chesson]

Ddp

Datagram Delivery Protocol [Wesley_Craig]

IdprCmtp

IDPR Control Message Transport Proto [Martha_Steenstrup]

TpPlusPlus

TP++ Transport Protocol [Dirk_Fromhein]

Il

IL Transport Protocol [Dave_Presotto]

Ipv6

IPv6 encapsulation [RFC2473]

Sdrp

Source Demand Routing Protocol [Deborah_Estrin]

IPv6RouteHeader

Routing Header for IPv6 [Steve_Deering]

IPv6FragmentationHeader

Fragment Header for IPv6 [Steve_Deering]

Idrp

Inter-Domain Routing Protocol [Sue_Hares]

Rsvp

Reservation Protocol [RFC2205][RFC3209][Bob_Braden]

Gre

Generic Routing Encapsulation [RFC2784][Tony_Li]

Dsr

Dynamic Source Routing Protocol [RFC4728]

Bna

BNA [Gary Salamon]

IPv6EncapSecurityPayload

Encap Security Payload [RFC4303]

IPv6AuthenticationHeader

Authentication Header [RFC4302]

Inlsp

Integrated Net Layer Security TUBA [K_Robert_Glenn]

Swipe

IP with Encryption (deprecated) [John_Ioannidis]

Narp

NBMA Address Resolution Protocol [RFC1735]

Mobile

IP Mobility [Charlie_Perkins]

Tlsp

Transport Layer Security Protocol using Kryptonet key management [Christer_Oberg]

Skip

SKIP [Tom_Markson]

IPv6Icmp

ICMP for IPv6 [RFC8200]

IPv6NoNextHeader

No Next Header for IPv6 [RFC8200]

IPv6DestinationOptions

Destination Options for IPv6 [RFC8200]

AnyHostInternalProtocol

any host internal protocol [Internet_Assigned_Numbers_Authority]

Cftp

CFTP [Forsdick, H., "CFTP", Network Message, Bolt Beranek and Newman, January 1982.][Harry_Forsdick]

AnyLocalNetwork

any local network [Internet_Assigned_Numbers_Authority]

SatExpak

SATNET and Backroom EXPAK [Steven_Blumenthal]

Krytolan

Kryptolan [Paul Liu]

Rvd

MIT Remote Virtual Disk Protocol [Michael_Greenwald]

Ippc

Internet Pluribus Packet Core [Steven_Blumenthal]

AnyDistributedFileSystem

any distributed file system [Internet_Assigned_Numbers_Authority]

SatMon

SATNET Monitoring [Steven_Blumenthal]

Visa

VISA Protocol [Gene_Tsudik]

Ipcv

Internet Packet Core Utility [Steven_Blumenthal]

Cpnx

Computer Protocol Network Executive [David Mittnacht]

Cphb

Computer Protocol Heart Beat [David Mittnacht]

Wsn

Wang Span Network [Victor Dafoulas]

Pvp

Packet Video Protocol [Steve_Casner]

BrSatMon

Backroom SATNET Monitoring [Steven_Blumenthal]

SunNd

SUN ND PROTOCOL-Temporary [William_Melohn]

WbMon

WIDEBAND Monitoring [Steven_Blumenthal]

WbExpak

WIDEBAND EXPAK [Steven_Blumenthal]

IsoIp

ISO Internet Protocol [Marshall_T_Rose]

Vmtp

VMTP [Dave_Cheriton]

SecureVmtp

SECURE-VMTP [Dave_Cheriton]

Vines

VINES [Brian Horn]

TtpOrIptm

Transaction Transport Protocol or Internet Protocol Traffic Manager [Jim_Stevens]

NsfnetIgp

NSFNET-IGP [Hans_Werner_Braun]

Dgp

Dissimilar Gateway Protocol [M/A-COM Government Systems, "Dissimilar Gateway Protocol Specification, Draft Version", Contract no. CS901145, November 16, 1987.][Mike_Little]

Tcf

TCF [Guillermo_A_Loyola]

Eigrp

EIGRP [RFC7868]

Ospfigp

OSPFIGP [RFC1583][RFC2328][RFC5340][John_Moy]

SpriteRpc

Sprite RPC Protocol [Welch, B., "The Sprite Remote Procedure Call System", Technical Report, UCB/Computer Science Dept., 86/302, University of California at Berkeley, June 1986.][Bruce Willins]

Larp

Locus Address Resolution Protocol [Brian Horn]

Mtp

Multicast Transport Protocol [Susie_Armstrong]

Ax25

AX.25 Frames [Brian_Kantor]

Ipip

IP-within-IP Encapsulation Protocol [John_Ioannidis]

Micp

Mobile Internetworking Control Pro. (deprecated) [John_Ioannidis]

SccSp

Semaphore Communications Sec. Pro. [Howard_Hart]

EtherIp

Ethernet-within-IP Encapsulation [RFC3378]

Encap

Encapsulation Header [RFC1241][Robert_Woodburn]

Gmtp

GMTP [[RXB5]]

Ifmp

Ipsilon Flow Management Protocol [Bob_Hinden][November 1995, 1997.]

Pnni

PNNI over IP [Ross_Callon]

Pim

Protocol Independent Multicast [RFC7761][Dino_Farinacci]

Aris

ARIS [Nancy_Feldman]

Scps

SCPS [Robert_Durst]

Qnx

QNX [Michael_Hunter]

ActiveNetworks

Active Networks [Bob_Braden]

IpComp

IP Payload Compression Protocol [RFC2393]

SitraNetworksProtocol

Sitara Networks Protocol [Manickam_R_Sridhar]

CompaqPeer

Compaq Peer Protocol [Victor_Volpe]

IpxInIp

IPX in IP [CJ_Lee]

Vrrp

Virtual Router Redundancy Protocol [RFC5798]

Pgm

PGM Reliable Transport Protocol [Tony_Speakman]

AnyZeroHopProtocol

any 0-hop protocol [Internet_Assigned_Numbers_Authority]

Layer2TunnelingProtocol

Layer Two Tunneling Protocol [RFC3931][Bernard_Aboba]

Ddx

D-II Data Exchange (DDX) [John_Worley]

Iatp

Interactive Agent Transfer Protocol [John_Murphy]

Stp

Schedule Transfer Protocol [Jean_Michel_Pittet]

Srp

SpectraLink Radio Protocol [Mark_Hamilton]

Uti

UTI [Peter_Lothberg]

SimpleMessageProtocol

Simple Message Protocol [Leif_Ekblad]

Sm

Simple Multicast Protocol (deprecated) [Jon_Crowcroft][draft-perlman-simple-multicast]

Ptp

Performance Transparency Protocol [Michael_Welzl]

IsisOverIpv4

ISIS over IPv4 [Tony_Przygienda]

Fire

FIRE [Criag_Partridge]

Crtp

Combat Radio Transport Protocol [Robert_Sautter]

Crudp

Combat Radio User Datagram [Robert_Sautter]

Sscopmce

SSCOPMCE [Kurt_Waber]

Iplt

IPLT [[Hollbach]]

Sps

Secure Packet Shield [Bill_McIntosh]

Pipe

Private IP Encapsulation within IP [Bernhard_Petri]

Sctp

Stream Control Transmission Protocol [Randall_R_Stewart]

Fc

Fibre Channel [Murali_Rajagopal][RFC6172]

RsvpE2eIgnore

RSVP-E2E-IGNORE [RFC3175]

MobilityHeader

MobilityHeader [RFC6275]

UdpLite

UDPLite [RFC3828]

MplsInIp

[RFC4023]

Manet

MANET Protocols [RFC5498]

Hip

Host Identity Protocol [RFC7401]

Shim6

Shim6 Protocol [RFC5533]

Wesp

Wrapped Encapsulating Security Payload [RFC5840]

Rohc

Robust Header Compression [RFC5858]

ExperimentalAndTesting0

Use for experimentation and testing

ExperimentalAndTesting1

Use for experimentation and testing

Implementations

impl IpTrafficClass

pub fn is_ipv6_ext_header_value(value: u8) -> bool

Returns true if the given id identifies an IPV6 extension header traffic class.