etherparse

Module icmpv4

Module containing ICMPv4 related types and constants.

Structs

RedirectHeader

TimestampMessage

A ICMPv4 timestamp or timestamp response message.

Enums

DestUnreachableHeader

“Destination Unreachable” ICMP header for IPv4 (without the invoking packet).

ParameterProblemHeader

The header of an ICMPv4 Parameter Problems (contents up to the offending ip header).

RedirectCode

Code value in an ICMPv4 Redirect message.

TimeExceededCode

Code values for ICMPv4 time exceeded message.

Constants

CODE_DST_UNREACH_FILTER_PROHIB

ICMP destination unreachable code for “Communication Administratively Prohibited” (defined in RFC 1812)

CODE_DST_UNREACH_HOST

ICMP destination unreachable code for “Host Unreachable” (defined in RFC 792)

CODE_DST_UNREACH_HOST_PRECEDENCE_VIOLATION

ICMP destination unreachable code for “Host Precedence Violation” (defined in RFC 1812)

CODE_DST_UNREACH_HOST_PROHIB

ICMP destination unreachable code for “Communication with Destination Host is Administratively Prohibited” (defined in RFC 1122)

CODE_DST_UNREACH_HOST_UNKNOWN

ICMP destination unreachable code for “Destination Host Unknown” (defined in RFC 1122)

CODE_DST_UNREACH_ISOLATED

ICMP destination unreachable code for “Source Host Isolated” (defined in RFC 1122)

CODE_DST_UNREACH_NEED_FRAG

ICMP destination unreachable code for “Fragmentation Needed and Don’t Fragment was Set” (defined in RFC 792)

CODE_DST_UNREACH_NET

ICMP destination unreachable code for “Net Unreachable” (defined in RFC 792)

CODE_DST_UNREACH_NET_PROHIB

ICMP destination unreachable code for “Communication with Destination Network is Administratively Prohibited” (defined in RFC 1122)

CODE_DST_UNREACH_NET_UNKNOWN

ICMP destination unreachable code for “Destination Network Unknown” (defined in RFC 1122)

CODE_DST_UNREACH_PORT

ICMP destination unreachable code for “Port Unreachable” (defined in RFC 792)

CODE_DST_UNREACH_PRECEDENCE_CUTOFF

ICMP destination unreachable code for “Precedence cutoff in effect” (defined in RFC 1812)

CODE_DST_UNREACH_PROTOCOL

ICMP destination unreachable code for “Protocol Unreachable” (defined in RFC 792)

CODE_DST_UNREACH_SOURCE_ROUTE_FAILED

ICMP destination unreachable code for “Source Route Failed” (defined in RFC 792)

CODE_DST_UNREACH_TOS_HOST

ICMP destination unreachable code for “Destination Host Unreachable for Type of Service” (defined in RFC 1122)

CODE_DST_UNREACH_TOS_NET

ICMP destination unreachable code for “Destination Network Unreachable for Type of Service” (defined in RFC 1122)

CODE_PARAMETER_PROBLEM_BAD_LENGTH

ICMPv4 “Parameter Problem” code value for “Bad Length”.

CODE_PARAMETER_PROBLEM_MISSING_REQUIRED_OPTION

ICMPv4 “Parameter Problem” code value for “Missing a Required Option”.

CODE_PARAMETER_PROBLEM_POINTER_INDICATES_ERROR

ICMPv4 “Parameter Problem” code value for “Pointer indicates the error”.

CODE_REDIRECT_FOR_HOST

ICMPv4 “Redirect” code value for “Redirect Datagram for the Host”.

CODE_REDIRECT_FOR_NETWORK

ICMPv4 “Redirect” code value for “Redirect Datagram for the Network (or subnet)”.

CODE_REDIRECT_TYPE_OF_SERVICE_AND_HOST

ICMPv4 “Redirect” code value for “Redirect Datagram for the Type of Service and Host”.

CODE_REDIRECT_TYPE_OF_SERVICE_AND_NETWORK

ICMPv4 “Redirect” code value for “Redirect Datagram for the Type of Service and Network”.

CODE_TIME_EXCEEDED_FRAG_REASSEMBLY_TIME_EXCEEDED

ICMPv4 “Time Exceeded” code value for “Fragment Reassembly Time Exceeded”.

CODE_TIME_EXCEEDED_TTL_EXCEEDED_IN_TRANSIT

ICMPv4 “Time Exceeded” code value for “Time to Live exceeded in Transit”.

TYPE_ADDRESS

ICMPv4 type value indicating a “Address Mask Request (Deprecated)” message (defined in in RFC 950, deprecated in RFC 6918).

TYPE_ADDRESSREPLY

ICMPv4 type value indicating a “Address Mask Reply (Deprecated)” message (defined in in RFC 950, deprecated in RFC 6918).

TYPE_ALTERNATE_HOST_ADDRESS

ICMPv4 type value indicating a “Alternate Host Address (Deprecated)” message (deprecated in RFC 6918).

TYPE_DEST_UNREACH

ICMPv4 type value indicating a “Destination Unreachable” message (defined in RFC 792).

TYPE_ECHO_REPLY

ICMPv4 type value indicating a “Echo Reply” message (defined in RFC 792).

TYPE_ECHO_REQUEST

ICMPv4 type value indicating a “Echo Request” message (defined in RFC 792).

TYPE_INFO_REPLY

ICMPv4 type value indicating a “Information Reply (Deprecated)” message (defined in in RFC 792, deprecated in RFC 6918).

TYPE_INFO_REQUEST

ICMPv4 type value indicating a “Information Request (Deprecated)” message (defined in in RFC 792, deprecated in RFC 6918).

TYPE_PARAMETER_PROBLEM

ICMPv4 type value indicating a “Parameter Problem” message (defined in RFC 792).

TYPE_REDIRECT

ICMPv4 type value indicating a “Redirect” message (defined in RFC 792).

TYPE_ROUTER_ADVERTISEMENT

ICMPv4 type value indicating a “Router Advertisement” message (defined in RFC 1256).

TYPE_ROUTER_SOLICITATION

ICMPv4 type value indicating a “Router Solicitation” message (defined in RFC 1256).

TYPE_SOURCE_QUENCH

ICMPv4 type value indicating a “Source Quench (Deprecated)” message (defined in in RFC 792, deprecated in RFC 6633).

TYPE_TIMESTAMP

ICMPv4 type value indicating a “Timestamp” message (defined in RFC 792).

TYPE_TIMESTAMP_REPLY

ICMPv4 type value indicating a “Timestamp Reply” message (defined in RFC 792).

TYPE_TIME_EXCEEDED

ICMPv4 type value indicating a “Time Exceeded” message (defined in RFC 792).