etherparse

Module ip_number

Constants for the ip protocol numbers for easy importing (e.g. use ip_number::*;).

The constants only exist for convenience so you can import them (use ip_number::*) without a need to write IpNumber:: in front of every constant.

You can access the underlying u8 value by using .0 and any u8 can be converted to an IpNumber:

use etherparse::{ip_number::TCP, IpNumber};

assert_eq!(TCP.0, 6);
assert_eq!(TCP, IpNumber(6));
let num: IpNumber = 6.into();
assert_eq!(TCP, num);

The list original values were copied from https://www.iana.org/assignments/protocol-numbers/protocol-numbers.xhtml

Constants

ACTIVE_NETWORKS

Active Networks [Bob_Braden]

ANY_DISTRIBUTED_FILE_SYSTEM

any distributed file system [Internet_Assigned_Numbers_Authority]

ANY_HOST_INTERNAL_PROTOCOL

any host internal protocol [Internet_Assigned_Numbers_Authority]

ANY_LOCAL_NETWORK

any local network [Internet_Assigned_Numbers_Authority]

ANY_ZERO_HOP_PROTOCOL

any 0-hop protocol [Internet_Assigned_Numbers_Authority]

ARGUS

ARGUS (deprecated) [Robert_W_Scheifler]

ARIS

ARIS [Nancy_Feldman]

AUTH

Authentication Header [RFC4302]

AUTHENTICATION_HEADER

Authentication Header [RFC4302]

AX25

AX.25 Frames [Brian_Kantor]

BBN_RCC_MON

BBN RCC Monitoring [Steve_Chipman]

BNA

BNA [Gary Salamon]

BR_SAT_MON

Backroom SATNET Monitoring [Steven_Blumenthal]

CBT

CBT [Tony_Ballardie]

CFTP

CFTP [Forsdick, H., “CFTP”, Network Message, Bolt Beranek and Newman, January 1982.][Harry_Forsdick]

CHAOS

Chaos [J_Noel_Chiappa]

COMPAQ_PEER

Compaq Peer Protocol [Victor_Volpe]

CPHB

Computer Protocol Heart Beat [David Mittnacht]

CPNX

Computer Protocol Network Executive [David Mittnacht]

CRTP

Combat Radio Transport Protocol [Robert_Sautter]

CRUDP

Combat Radio User Datagram [Robert_Sautter]

DCCP

Datagram Congestion Control Protocol [RFC4340]

DCN_MEAS

DCN Measurement Subsystems [David_Mills]

DDP

Datagram Delivery Protocol [Wesley_Craig]

DDX

D-II Data Exchange (DDX) [John_Worley]

DGP

Dissimilar Gateway Protocol [M/A-COM Government Systems, “Dissimilar Gateway Protocol Specification, Draft Version”, Contract no. CS901145, November 16, 1987.][Mike_Little]

DSR

Dynamic Source Routing Protocol [RFC4728]

EGP

Exterior Gateway Protocol [RFC888] [David_Mills]

EIGRP

EIGRP [RFC7868]

EMCON

EMCON [mystery contact]

ENCAP

Encapsulation Header [RFC1241][Robert_Woodburn]

ENCAPSULATING_SECURITY_PAYLOAD

Encapsulating Security Payload [RFC4303]

ENCAP_SEC

Encapsulating Security Payload [RFC4303]

ETHER_IP

Ethernet-within-IP Encapsulation [RFC3378]

EXP0

Use for experimentation and testing

EXP1

Use for experimentation and testing

EXPERIMENTAL_AND_TESTING_0

Use for experimentation and testing

EXPERIMENTAL_AND_TESTING_1

Use for experimentation and testing

FC

Fibre Channel [Murali_Rajagopal][RFC6172]

FIRE

FIRE [Criag_Partridge]

GGP

Gateway-to-Gateway [RFC823]

GMTP

GMTP [[RXB5]]

GRE

Generic Routing Encapsulation [RFC2784][Tony_Li]

HIP

Host Identity Protocol [RFC7401]

HMP

Host Monitoring [RFC869][Bob_Hinden]

IATP

Interactive Agent Transfer Protocol [John_Murphy]

ICMP

Internet Control Message [RFC792]

IDPR

Inter-Domain Policy Routing Protocol [Martha_Steenstrup]

IDPR_CMTP

IDPR Control Message Transport Proto [Martha_Steenstrup]

IDRP

Inter-Domain Routing Protocol [Sue_Hares]

IFMP

Ipsilon Flow Management Protocol [Bob_Hinden][November 1995, 1997.]

IGMP

Internet Group Management [RFC1112]

IGP

any private interior gateway (used by Cisco for their IGRP) [Internet_Assigned_Numbers_Authority]

IL

IL Transport Protocol [Dave_Presotto]

INLSP

Integrated Net Layer Security TUBA [K_Robert_Glenn]

IPCV

Internet Packet Core Utility [Steven_Blumenthal]

IPIP

IP-within-IP Encapsulation Protocol [John_Ioannidis]

IPLT

IPLT [[Hollbach]]

IPPC

Internet Pluribus Packet Core [Steven_Blumenthal]

IPV4

IPv4 encapsulation [RFC2003]

IPV6

IPv6 encapsulation [RFC2473]

IPV6_DESTINATION_OPTIONS

Destination Options for IPv6 [RFC8200]

IPV6_DEST_OPTIONS

Destination Options for IPv6 [RFC8200]

IPV6_FRAG

Fragment Header for IPv6 [Steve_Deering]

IPV6_FRAGMENTATION_HEADER

Fragment Header for IPv6 [Steve_Deering]

IPV6_HEADER_HOP_BY_HOP

IPv6 Hop-by-Hop Option [RFC8200]

IPV6_HOP_BY_HOP

IPv6 Hop-by-Hop Option [RFC8200]

IPV6_ICMP

IPv6 ICMP next-header type [RFC4443]

IPV6_NO_NEXT_HEADER

No Next Header for IPv6 [RFC8200]

IPV6_ROUTE

Routing Header for IPv6 [Steve_Deering]

IPV6_ROUTE_HEADER

Routing Header for IPv6 [Steve_Deering]

IPX_IN_IP

IPX in IP [CJ_Lee]

IP_COMP

IP Payload Compression Protocol [RFC2393]

IRTP

Internet Reliable Transaction [RFC938] [Trudy_Miller]

ISIS_OVER_IPV4

ISIS over IPv4 [Tony_Przygienda]

ISO_IP

ISO Internet Protocol [Marshall_T_Rose]

ISO_TP4

ISO Transport Protocol Class 4 [RFC905] [mystery contact]

KRYTOLAN

Kryptolan [Paul Liu]

LARP

Locus Address Resolution Protocol [Brian Horn]

LAYER2_TUNNELING_PROTOCOL

Layer Two Tunneling Protocol [RFC3931][Bernard_Aboba]

LEAF1

Leaf-1 [Barry_Boehm]

LEAF2

Leaf-2 [Barry_Boehm]

MANET

MANET Protocols [RFC5498]

MERIT_INP

MERIT Internodal Protocol [Hans_Werner_Braun]

MFE_NSP

MFE Network Services Protocol [Shuttleworth, B., “A Documentary of MFENet, a National Computer Network”, UCRL-52317, Lawrence Livermore Labs, Livermore, California, June 1977.] [Barry_Howard]

MICP

Mobile Internetworking Control Pro. (deprecated) [John_Ioannidis]

MOBILE

IP Mobility [Charlie_Perkins]

MOBILITY

MobilityHeader [RFC6275]

MOBILITY_HEADER

MobilityHeader [RFC6275]

MPLS_IN_IP

[RFC4023]

MTP

Multicast Transport Protocol [Susie_Armstrong]

MUX

Multiplexing [Cohen, D. and J. Postel, “Multiplexing Protocol”, IEN 90, USC/Information Sciences Institute, May 1979.][Jon_Postel]

NARP

NBMA Address Resolution Protocol [RFC1735]

NET_BLT

Bulk Data Transfer Protocol [RFC969] [David_Clark]

NSFNET_IGP

NSFNET-IGP [Hans_Werner_Braun]

NVP_II

Network Voice Protocol [RFC741][Steve_Casner]

OSPFIGP

OSPFIGP [RFC1583][RFC2328][RFC5340][John_Moy]

PGM

PGM Reliable Transport Protocol [Tony_Speakman]

PIM

Protocol Independent Multicast [RFC7761][Dino_Farinacci]

PIPE

Private IP Encapsulation within IP [Bernhard_Petri]

PNNI

PNNI over IP [Ross_Callon]

PRM

Packet Radio Measurement [Zaw_Sing_Su]

PTP

Performance Transparency Protocol [Michael_Welzl]

PUP

PUP

PVP

Packet Video Protocol [Steve_Casner]

QNX

QNX [Michael_Hunter]

RDP

Reliable Data Protocol [RFC908] [Bob_Hinden]

ROHC

Robust Header Compression [RFC5858]

RSVP

Reservation Protocol [RFC2205][RFC3209][Bob_Braden]

RSVP_E2E_IGNORE

RSVP-E2E-IGNORE [RFC3175]

RVD

MIT Remote Virtual Disk Protocol [Michael_Greenwald]

SAT_EXPAK

SATNET and Backroom EXPAK [Steven_Blumenthal]

SAT_MON

SATNET Monitoring [Steven_Blumenthal]

SCC_SP

Semaphore Communications Sec. Pro. [Howard_Hart]

SCPS

SCPS [Robert_Durst]

SCTP

Stream Control Transmission Protocol [Randall_R_Stewart]

SDRP

Source Demand Routing Protocol [Deborah_Estrin]

SECURE_VMTP

SECURE-VMTP [Dave_Cheriton]

SHIM6

Shim6 Protocol [RFC5533]

SIMPLE_MESSAGE_PROTOCOL

Simple Message Protocol [Leif_Ekblad]

SITRA_NETWORKS_PROTOCOL

Sitara Networks Protocol [Manickam_R_Sridhar]

SKIP

SKIP [Tom_Markson]

SM

Simple Multicast Protocol (deprecated) [Jon_Crowcroft][draft-perlman-simple-multicast]

SPRITE_RPC

Sprite RPC Protocol [Welch, B., “The Sprite Remote Procedure Call System”, Technical Report, UCB/Computer Science Dept., 86/302, University of California at Berkeley, June 1986.][Bruce Willins]

SPS

Secure Packet Shield [Bill_McIntosh]

SRP

SpectraLink Radio Protocol [Mark_Hamilton]

SSCOPMCE

SSCOPMCE [Kurt_Waber]

STP

Schedule Transfer Protocol [Jean_Michel_Pittet]

STREAM

Stream [RFC1190] [RFC1819]

SUN_ND

SUN ND PROTOCOL-Temporary [William_Melohn]

SWIPE

IP with Encryption (deprecated) [John_Ioannidis]

TCF

TCF [Guillermo_A_Loyola]

TCP

Transmission Control [RFC793]

THIRD_PARTY_CONNECT_PROTOCOL

Third Party Connect Protocol [Stuart_A_Friedberg]

TLSP

Transport Layer Security Protocol using Kryptonet key management [Christer_Oberg]

TP_PLUS_PLUS

TP++ Transport Protocol [Dirk_Fromhein]

TRUNK1

Trunk-1 [Barry_Boehm]

TRUNK2

Trunk-2 [Barry_Boehm]

TTP_OR_IPTM

Transaction Transport Protocol or Internet Protocol Traffic Manager [Jim_Stevens]

UDP

User Datagram [RFC768] [Jon_Postel]

UDP_LITE

UDPLite [RFC3828]

UTI

UTI [Peter_Lothberg]

VINES

VINES [Brian Horn]

VISA

VISA Protocol [Gene_Tsudik]

VMTP

VMTP [Dave_Cheriton]

VRRP

Virtual Router Redundancy Protocol [RFC5798]

WB_EXPAK

WIDEBAND EXPAK [Steven_Blumenthal]

WB_MON

WIDEBAND Monitoring [Steven_Blumenthal]

WESP

Wrapped Encapsulating Security Payload [RFC5840]

WSN

Wang Span Network [Victor Dafoulas]

XNET

Cross Net Debugger [Haverty, J., “XNET Formats for Internet Protocol Version 4”, IEN 158, October 1980.][Jack_Haverty]

XNS_IDP

XEROX NS IDP

XTP

XTP [Greg_Chesson]