etherparse

Struct Icmpv6Header

pub struct Icmpv6Header {
    pub icmp_type: Icmpv6Type,
    pub checksum: u16,
}

The statically sized data at the start of an ICMPv6 packet (at least the first 8 bytes of an ICMPv6 packet).

Fields
icmp_type: Icmpv6Type

Type & type specific values & code.

checksum: u16

Checksum in the ICMPv6 header.

Implementations

impl Icmpv6Header

pub const MIN_LEN: usize = 8usize

Minimum number of bytes an ICMP header needs to have.

Note that minimum size can be larger depending on the type and code.

pub const MIN_SERIALIZED_SIZE: usize = 8usize

๐Deprecated since 0.14.0: Please use Icmpv6Header::MIN_LEN instead

Deprecated, use Icmpv6Header::MIN_LEN instead.

pub const MAX_LEN: usize = 40usize

Maximum number of bytes/octets an Icmpv6Header takes up in serialized form.

Currently this number is determined by the biggest planned ICMPv6 header type, which is currently the โNeighbor Discovery Protocolโ โRedirectโ message.

pub const MAX_SERIALIZED_SIZE: usize = 40usize

๐Deprecated since 0.14.0: Please use Icmpv6Header::MAX_LEN instead

Deprecated, use Icmpv6Header::MAX_LEN instead.

pub fn new(icmp_type: Icmpv6Type) -> Icmpv6Header

Setups a new header with the checksum being set to 0.

pub fn with_checksum( icmp_type: Icmpv6Type, source_ip: [u8; 16], destination_ip: [u8; 16], payload: &[u8], ) -> Result<Icmpv6Header, ValueTooBigError<usize>>

Creates a Icmpv6Header with a checksum calculated based on the given payload & ip addresses from the IPv6 header.

pub fn from_slice(slice: &[u8]) -> Result<(Icmpv6Header, &[u8]), LenError>

Reads an icmp6 header from a slice directly and returns a tuple containing the resulting header & unused part of the slice.

pub fn read<T: Read + Sized>(reader: &mut T) -> Result<Icmpv6Header, Error>

Read a ICMPv6 header from the given reader

pub fn write<T: Write + Sized>(&self, writer: &mut T) -> Result<(), Error>

Write the ICMPv6 header to the given writer.

pub fn header_len(&self) -> usize

Serialized length of the header in bytes/octets.

Note that this size is not the size of the entire ICMPv6 packet but only the header.

pub fn fixed_payload_size(&self) -> Option<usize>

If the ICMP type has a fixed size returns the number of bytes that should be present after the header of this type.

pub fn update_checksum( &mut self, source_ip: [u8; 16], destination_ip: [u8; 16], payload: &[u8], ) -> Result<(), ValueTooBigError<usize>>

Updates the checksum of the header.

pub fn to_bytes(&self) -> ArrayVec<u8, { Icmpv6Header::MAX_LEN }>

Returns the header on the wire bytes.