etherparse

Struct IpNumber

pub struct IpNumber(pub u8);

Identifiers for the next_header field in ipv6 headers and protocol field in ipv4 headers.

You can access the underlying u8 value by using .0 and any u8 can be converted to an IpNumber:

use etherparse::IpNumber;

assert_eq!(IpNumber::TCP.0, 6);
assert_eq!(IpNumber::TCP, IpNumber(6));

// convert to IpNumber using the from & into trait
let ip_num: IpNumber = 6.into();
assert_eq!(IpNumber::TCP, ip_num);

// convert to u8 using the from & into trait
let num: u8 = IpNumber::TCP.into();
assert_eq!(6, num);

The constants are also defined in the ip_number module so they can be used without the need to write IpNumber:: in front of them:

use etherparse::{ip_number::TCP, IpNumber};

assert_eq!(TCP, IpNumber::TCP);

The list original values were copied from https://www.iana.org/assignments/protocol-numbers/protocol-numbers.xhtml

Tuple Fields
0: u8
Implementations

impl IpNumber

pub const IPV6_HEADER_HOP_BY_HOP: IpNumber

IPv6 Hop-by-Hop Option [RFC8200]

pub const ICMP: IpNumber

Internet Control Message [RFC792]

pub const IGMP: IpNumber

Internet Group Management [RFC1112]

pub const GGP: IpNumber

Gateway-to-Gateway [RFC823]

pub const IPV4: IpNumber

IPv4 encapsulation [RFC2003]

pub const STREAM: IpNumber

Stream [RFC1190] [RFC1819]

pub const TCP: IpNumber

Transmission Control [RFC793]

pub const CBT: IpNumber

CBT [Tony_Ballardie]

pub const EGP: IpNumber

Exterior Gateway Protocol [RFC888] [David_Mills]

pub const IGP: IpNumber

any private interior gateway (used by Cisco for their IGRP) [Internet_Assigned_Numbers_Authority]

pub const BBN_RCC_MON: IpNumber

BBN RCC Monitoring [Steve_Chipman]

pub const NVP_II: IpNumber

Network Voice Protocol [RFC741][Steve_Casner]

pub const PUP: IpNumber

PUP

pub const ARGUS: IpNumber

ARGUS (deprecated) [Robert_W_Scheifler]

pub const EMCON: IpNumber

EMCON [mystery contact]

pub const XNET: IpNumber

Cross Net Debugger [Haverty, J., “XNET Formats for Internet Protocol Version 4”, IEN 158, October 1980.][Jack_Haverty]

pub const CHAOS: IpNumber

Chaos [J_Noel_Chiappa]

pub const UDP: IpNumber

User Datagram [RFC768][Jon_Postel]

pub const MUX: IpNumber

Multiplexing [Cohen, D. and J. Postel, “Multiplexing Protocol”, IEN 90, USC/Information Sciences Institute, May 1979.][Jon_Postel]

pub const DCN_MEAS: IpNumber

DCN Measurement Subsystems [David_Mills]

pub const HMP: IpNumber

Host Monitoring [RFC869][Bob_Hinden]

pub const PRM: IpNumber

Packet Radio Measurement [Zaw_Sing_Su]

pub const XNS_IDP: IpNumber

XEROX NS IDP

pub const TRUNK1: IpNumber

Trunk-1 [Barry_Boehm]

pub const TRUNK2: IpNumber

Trunk-2 [Barry_Boehm]

pub const LEAF1: IpNumber

Leaf-1 [Barry_Boehm]

pub const LEAF2: IpNumber

Leaf-2 [Barry_Boehm]

pub const RDP: IpNumber

Reliable Data Protocol [RFC908] [Bob_Hinden]

pub const IRTP: IpNumber

Internet Reliable Transaction [RFC938] [Trudy_Miller]

pub const ISO_TP4: IpNumber

ISO Transport Protocol Class 4 [RFC905] [mystery contact]

pub const NET_BLT: IpNumber

Bulk Data Transfer Protocol [RFC969] [David_Clark]

pub const MFE_NSP: IpNumber

MFE Network Services Protocol [Shuttleworth, B., “A Documentary of MFENet, a National Computer Network”, UCRL-52317, Lawrence Livermore Labs, Livermore, California, June 1977.] [Barry_Howard]

pub const MERIT_INP: IpNumber

MERIT Internodal Protocol [Hans_Werner_Braun]

pub const DCCP: IpNumber

Datagram Congestion Control Protocol [RFC4340]

pub const THIRD_PARTY_CONNECT_PROTOCOL: IpNumber

Third Party Connect Protocol [Stuart_A_Friedberg]

pub const IDPR: IpNumber

Inter-Domain Policy Routing Protocol [Martha_Steenstrup]

pub const XTP: IpNumber

XTP [Greg_Chesson]

pub const DDP: IpNumber

Datagram Delivery Protocol [Wesley_Craig]

pub const IDPR_CMTP: IpNumber

IDPR Control Message Transport Proto [Martha_Steenstrup]

pub const TP_PLUS_PLUS: IpNumber

TP++ Transport Protocol [Dirk_Fromhein]

pub const IL: IpNumber

IL Transport Protocol [Dave_Presotto]

pub const IPV6: IpNumber

IPv6 encapsulation [RFC2473]

pub const SDRP: IpNumber

Source Demand Routing Protocol [Deborah_Estrin]

pub const IPV6_ROUTE_HEADER: IpNumber

Routing Header for IPv6 [Steve_Deering]

pub const IPV6_FRAGMENTATION_HEADER: IpNumber

Fragment Header for IPv6 [Steve_Deering]

pub const IDRP: IpNumber

Inter-Domain Routing Protocol [Sue_Hares]

pub const RSVP: IpNumber

Reservation Protocol [RFC2205][RFC3209][Bob_Braden]

pub const GRE: IpNumber

Generic Routing Encapsulation [RFC2784][Tony_Li]

pub const DSR: IpNumber

Dynamic Source Routing Protocol [RFC4728]

pub const BNA: IpNumber

BNA [Gary Salamon]

pub const ENCAPSULATING_SECURITY_PAYLOAD: IpNumber

Encapsulating Security Payload [RFC4303]

pub const AUTHENTICATION_HEADER: IpNumber

Authentication Header [RFC4302]

pub const INLSP: IpNumber

Integrated Net Layer Security TUBA [K_Robert_Glenn]

pub const SWIPE: IpNumber

IP with Encryption (deprecated) [John_Ioannidis]

pub const NARP: IpNumber

NBMA Address Resolution Protocol [RFC1735]

pub const MOBILE: IpNumber

IP Mobility [Charlie_Perkins]

pub const TLSP: IpNumber

Transport Layer Security Protocol using Kryptonet key management [Christer_Oberg]

pub const SKIP: IpNumber

SKIP [Tom_Markson]

pub const IPV6_ICMP: IpNumber

ICMP for IPv6 [RFC8200]

pub const IPV6_NO_NEXT_HEADER: IpNumber

No Next Header for IPv6 [RFC8200]

pub const IPV6_DESTINATION_OPTIONS: IpNumber

Destination Options for IPv6 [RFC8200]

pub const ANY_HOST_INTERNAL_PROTOCOL: IpNumber

any host internal protocol [Internet_Assigned_Numbers_Authority]

pub const CFTP: IpNumber

CFTP [Forsdick, H., “CFTP”, Network Message, Bolt Beranek and Newman, January 1982.][Harry_Forsdick]

pub const ANY_LOCAL_NETWORK: IpNumber

any local network [Internet_Assigned_Numbers_Authority]

pub const SAT_EXPAK: IpNumber

SATNET and Backroom EXPAK [Steven_Blumenthal]

pub const KRYTOLAN: IpNumber

Kryptolan [Paul Liu]

pub const RVD: IpNumber

MIT Remote Virtual Disk Protocol [Michael_Greenwald]

pub const IPPC: IpNumber

Internet Pluribus Packet Core [Steven_Blumenthal]

pub const ANY_DISTRIBUTED_FILE_SYSTEM: IpNumber

any distributed file system [Internet_Assigned_Numbers_Authority]

pub const SAT_MON: IpNumber

SATNET Monitoring [Steven_Blumenthal]

pub const VISA: IpNumber

VISA Protocol [Gene_Tsudik]

pub const IPCV: IpNumber

Internet Packet Core Utility [Steven_Blumenthal]

pub const CPNX: IpNumber

Computer Protocol Network Executive [David Mittnacht]

pub const CPHB: IpNumber

Computer Protocol Heart Beat [David Mittnacht]

pub const WSN: IpNumber

Wang Span Network [Victor Dafoulas]

pub const PVP: IpNumber

Packet Video Protocol [Steve_Casner]

pub const BR_SAT_MON: IpNumber

Backroom SATNET Monitoring [Steven_Blumenthal]

pub const SUN_ND: IpNumber

SUN ND PROTOCOL-Temporary [William_Melohn]

pub const WB_MON: IpNumber

WIDEBAND Monitoring [Steven_Blumenthal]

pub const WB_EXPAK: IpNumber

WIDEBAND EXPAK [Steven_Blumenthal]

pub const ISO_IP: IpNumber

ISO Internet Protocol [Marshall_T_Rose]

pub const VMTP: IpNumber

VMTP [Dave_Cheriton]

pub const SECURE_VMTP: IpNumber

SECURE-VMTP [Dave_Cheriton]

pub const VINES: IpNumber

VINES [Brian Horn]

pub const TTP_OR_IPTM: IpNumber

Transaction Transport Protocol or Internet Protocol Traffic Manager [Jim_Stevens]

pub const NSFNET_IGP: IpNumber

NSFNET-IGP [Hans_Werner_Braun]

pub const DGP: IpNumber

Dissimilar Gateway Protocol [M/A-COM Government Systems, “Dissimilar Gateway Protocol Specification, Draft Version”, Contract no. CS901145, November 16, 1987.][Mike_Little]

pub const TCF: IpNumber

TCF [Guillermo_A_Loyola]

pub const EIGRP: IpNumber

EIGRP [RFC7868]

pub const OSPFIGP: IpNumber

OSPFIGP [RFC1583][RFC2328][RFC5340][John_Moy]

pub const SPRITE_RPC: IpNumber

Sprite RPC Protocol [Welch, B., “The Sprite Remote Procedure Call System”, Technical Report, UCB/Computer Science Dept., 86/302, University of California at Berkeley, June 1986.][Bruce Willins]

pub const LARP: IpNumber

Locus Address Resolution Protocol [Brian Horn]

pub const MTP: IpNumber

Multicast Transport Protocol [Susie_Armstrong]

pub const AX25: IpNumber

AX.25 Frames [Brian_Kantor]

pub const IPIP: IpNumber

IP-within-IP Encapsulation Protocol [John_Ioannidis]

pub const MICP: IpNumber

Mobile Internetworking Control Pro. (deprecated) [John_Ioannidis]

pub const SCC_SP: IpNumber

Semaphore Communications Sec. Pro. [Howard_Hart]

pub const ETHER_IP: IpNumber

Ethernet-within-IP Encapsulation [RFC3378]

pub const ENCAP: IpNumber

Encapsulation Header [RFC1241][Robert_Woodburn]

pub const GMTP: IpNumber

GMTP [[RXB5]]

pub const IFMP: IpNumber

Ipsilon Flow Management Protocol [Bob_Hinden][November 1995, 1997.]

pub const PNNI: IpNumber

PNNI over IP [Ross_Callon]

pub const PIM: IpNumber

Protocol Independent Multicast [RFC7761][Dino_Farinacci]

pub const ARIS: IpNumber

ARIS [Nancy_Feldman]

pub const SCPS: IpNumber

SCPS [Robert_Durst]

pub const QNX: IpNumber

QNX [Michael_Hunter]

pub const ACTIVE_NETWORKS: IpNumber

Active Networks [Bob_Braden]

pub const IP_COMP: IpNumber

IP Payload Compression Protocol [RFC2393]

pub const SITRA_NETWORKS_PROTOCOL: IpNumber

Sitara Networks Protocol [Manickam_R_Sridhar]

pub const COMPAQ_PEER: IpNumber

Compaq Peer Protocol [Victor_Volpe]

pub const IPX_IN_IP: IpNumber

IPX in IP [CJ_Lee]

pub const VRRP: IpNumber

Virtual Router Redundancy Protocol [RFC5798]

pub const PGM: IpNumber

PGM Reliable Transport Protocol [Tony_Speakman]

pub const ANY_ZERO_HOP_PROTOCOL: IpNumber

any 0-hop protocol [Internet_Assigned_Numbers_Authority]

pub const LAYER2_TUNNELING_PROTOCOL: IpNumber

Layer Two Tunneling Protocol [RFC3931][Bernard_Aboba]

pub const DDX: IpNumber

D-II Data Exchange (DDX) [John_Worley]

pub const IATP: IpNumber

Interactive Agent Transfer Protocol [John_Murphy]

pub const STP: IpNumber

Schedule Transfer Protocol [Jean_Michel_Pittet]

pub const SRP: IpNumber

SpectraLink Radio Protocol [Mark_Hamilton]

pub const UTI: IpNumber

UTI [Peter_Lothberg]

pub const SIMPLE_MESSAGE_PROTOCOL: IpNumber

Simple Message Protocol [Leif_Ekblad]

pub const SM: IpNumber

Simple Multicast Protocol (deprecated) [Jon_Crowcroft][draft-perlman-simple-multicast]

pub const PTP: IpNumber

Performance Transparency Protocol [Michael_Welzl]

pub const ISIS_OVER_IPV4: IpNumber

ISIS over IPv4 [Tony_Przygienda]

pub const FIRE: IpNumber

FIRE [Criag_Partridge]

pub const CRTP: IpNumber

Combat Radio Transport Protocol [Robert_Sautter]

pub const CRUDP: IpNumber

Combat Radio User Datagram [Robert_Sautter]

pub const SSCOPMCE: IpNumber

SSCOPMCE [Kurt_Waber]

pub const IPLT: IpNumber

IPLT [[Hollbach]]

pub const SPS: IpNumber

Secure Packet Shield [Bill_McIntosh]

pub const PIPE: IpNumber

Private IP Encapsulation within IP [Bernhard_Petri]

pub const SCTP: IpNumber

Stream Control Transmission Protocol [Randall_R_Stewart]

pub const FC: IpNumber

Fibre Channel [Murali_Rajagopal][RFC6172]

pub const RSVP_E2E_IGNORE: IpNumber

RSVP-E2E-IGNORE [RFC3175]

pub const MOBILITY_HEADER: IpNumber

MobilityHeader [RFC6275]

pub const UDP_LITE: IpNumber

UDPLite [RFC3828]

pub const MPLS_IN_IP: IpNumber

[RFC4023]

pub const MANET: IpNumber

MANET Protocols [RFC5498]

pub const HIP: IpNumber

Host Identity Protocol [RFC7401]

pub const SHIM6: IpNumber

Shim6 Protocol [RFC5533]

pub const WESP: IpNumber

Wrapped Encapsulating Security Payload [RFC5840]

pub const ROHC: IpNumber

Robust Header Compression [RFC5858]

pub const EXPERIMENTAL_AND_TESTING_0: IpNumber

Use for experimentation and testing

pub const EXPERIMENTAL_AND_TESTING_1: IpNumber

Use for experimentation and testing

impl IpNumber

pub fn is_ipv6_ext_header_value(self) -> bool

Returns true if the given number is the internet number of an IPV6 extension header.

pub fn keyword_str(self) -> Option<&'static str>

Returns the “keyword” string if known. Usually this is the abbreviation of the protocol.

Example

use etherparse::IpNumber;

assert_eq!(IpNumber::UDP.keyword_str(), Some("UDP"));

// Unassigned values return None
assert_eq!(IpNumber(145).keyword_str(), None);

Data Source

The strings were copied from https://www.iana.org/assignments/protocol-numbers/protocol-numbers.xhtml on 2023-04-11.

pub fn protocol_str(self) -> Option<&'static str>

Returns the “protocol” string if known. Usually this the non abbreviated name of the protocol.

Example

use etherparse::IpNumber;

assert_eq!(IpNumber::UDP.protocol_str(), Some("User Datagram"));

// Unassigned values return None
assert_eq!(IpNumber(145).protocol_str(), None);

Data Source

The string was copied from https://www.iana.org/assignments/protocol-numbers/protocol-numbers.xhtml on 2023-04-11.