Struct etherparse::Ipv6HeaderSlice

pub struct Ipv6HeaderSlice<'a> { /* fields omitted */ }

A slice containing an ipv6 header of a network package.

Implementations

impl<'a> Ipv6HeaderSlice<'a>

pub fn from_slice(slice: &'a [u8]) -> Result<Ipv6HeaderSlice<'a>, ReadError>

Creates a slice containing an ipv6 header (without header extensions).

pub fn slice(&self) -> &'a [u8]

Returns the slice containing the ipv6 header

pub fn version(&self) -> u8

Read the "version" field from the slice (should be 6).

pub fn traffic_class(&self) -> u8

Read the "traffic class" field from the slice.

pub fn flow_label(&self) -> u32

Read the "flow label" field from the slice.

pub fn payload_length(&self) -> u16

Read the "payload length" field from the slice. The length should contain the length of all extension headers and payload.

pub fn next_header(&self) -> u8

Read the "next header" field from the slice. The next header value specifies what the next header or transport layer protocol is (see IpTrafficClass for a definitions of ids).

pub fn hop_limit(&self) -> u8

Read the "hop limit" field from the slice. The hop limit specifies the number of hops the packet can take before it is discarded.

pub fn source(&self) -> &'a [u8]

Returns a slice containing the IPv6 source address.

pub fn source_addr(&self) -> Ipv6Addr

Return the ipv6 source address as an std::net::Ipv6Addr

pub fn destination(&self) -> &'a [u8]

Returns a slice containing the IPv6 destination address.

pub fn destination_addr(&self) -> Ipv6Addr

Return the ipv6 destination address as an std::net::Ipv6Addr

pub fn to_header(&self) -> Ipv6Header

Decode all the fields and copy the results to a Ipv6Header struct