etherparse

Type Alias IpAuthenticationHeader

pub type IpAuthenticationHeader = IpAuthHeader;

๐Deprecated since 0.14.0: Please use the type IpAuthHeader instead

Deprecated use IpAuthHeader instead.

Aliased Type

struct IpAuthenticationHeader {
    pub next_header: IpNumber,
    pub spi: u32,
    pub sequence_number: u32,
    /* private fields */
}

Fields
next_header: IpNumber

IP protocol number specifying the next header or transport layer protocol.

See IpNumber or ip_number for a definition of the known values.

spi: u32

Security Parameters Index

sequence_number: u32

This unsigned 32-bit field contains a counter value that increases by one for each packet sent.