etherparse

Type Alias ReadError

pub type ReadError = ReadError;

๐Deprecated since 0.14.0: Please use the type err::ReadError instead or use the specific error type returned by operation you are using.

Deprecated use err::ReadError instead or use the specific error type returned by operation you are using.

Aliased Type

enum ReadError {
    Io(Error),
    Len(LenError),
    DoubleVlan(HeaderError),
    Ip(HeaderError),
    IpAuth(HeaderError),
    Ipv4(HeaderError),
    Ipv6(HeaderError),
    Ipv6Exts(HeaderError),
    Tcp(HeaderError),
}

Variants

Io(Error)

IO error was encountered while reading header or expected packet contents.

Len(LenError)

Error when parsing had to be aborted because of a length error (usually not enough data being available).

DoubleVlan(HeaderError)

Error while parsing a double vlan header.

Ip(HeaderError)

Error while parsing a IP header.

IpAuth(HeaderError)

Error while parsing a IP authentication header.

Ipv4(HeaderError)

Error while parsing a IPv4 header.

Ipv6(HeaderError)

Error while parsing a IPv6 header.

Ipv6Exts(HeaderError)

Error while parsing a IPv6 extension header.

Tcp(HeaderError)

Error while parsing a TCP extension header.